manage: remove clean_tmp.file_in_use
The files in TEMP_DIR are used in two stages:

- created by get_bulk_archive
- served by apache via X-Sendfile

If the file is removed while being served by apache, it does not matter
because the file descriptor will still be held by the process and the
connection will not be interrupted before the read completes.

The file_in_use will be useful if clean_tmp runs while the file is being
created by get_bulk_archive. It however misses a window between the
completion of get_bulk_archive (the file is closed and wont show via
file_in_use) and when apache opens it so it is served.

file_in_use is removed because it does not resolves the race between
get_bulk_archive+apache and clean_tmp in all cases.