OSSEC rules: Remove spurious --MARK-- ossec-keepalive alert
This alert is sent from the OSSEC agent to the manager and should
be filtered out by OSSEC but occasionally produces an alert.

This commit uses both the match and program_name OSSEC options to select
these events, marks them as Level 0, and ensures that no email
alerts are produced in the future.