OSSEC Rules: Remove syscheck alert when bulk tmp file deleted
When bulk downloads occur on the SecureDrop journalist interface,
a temporary file is created in /var/lib/securedrop/tmp. When
this file is cleaned up, it triggers an integrity checking alert.
If SecureDrop is getting heavy use, this can fire many times per
day.

This commit modifies the syscheck options on the app server
to exclude any files named tmp_securedrop_bulk_* in
/var/lib/securedrop/tmp from integrity checking (via syscheck).